Feature backend securitzation posada al dia respecte dev

+++ metari-app/src/components/CategoryList.tsx
@@ -37,11 +37,11 @@ export function CategoryList({ categories, setter, filteredCategory, setFiltered
 
     <>
       <div className="metaList mt-4">
-        <div className="titolComponent text-center my-2">Llista de Categories</div>
-        <hr className="m-0" />
+        <div className="titolComponent my-2 "><i className="bi bi-grid-fill text-primary"></i>Llista de Categories</div>
+        {/* <hr className="m-0" /> */}
 
         <div className="inline">
-          <ul className=" ps-2  m-0  py-2">
+          <ul className=" ps-3  m-0  pb-2">
             {categories.map((category) => (
               <li key={category.id} className="m-0 p-0" >
                 <div className={`metaEntry mt-1 me-3 ps-2 ${openEntitySelected === category.id ? "mb-0" : "mb-1"}`}
@@ -72,9 +72,17 @@ export function CategoryList({ categories, setter, filteredCategory, setFiltered
                     }
                     {canEdit &&
                       <button className="  btn btn-danger p-1   "
+                      title="Eliminar"
                         onClick={async (event) => {
                           event.stopPropagation()
-                          await deleteCategory(category.id)
+                          try {
+                            if (!confirm("Estàs segur que el vols eliminar?")) return;
+                            await deleteCategory(category.id)
+                            setter(prev => prev.filter(prevCategory => prevCategory.id !== category.id))
+                            alert("Categoria eliminada correctament")
+                          } catch (error) {
+                            alert("Error en eliminar la categoria")
+                          }
                         }}>X</button>
                     }
                   </div>
@@ -82,8 +90,8 @@ export function CategoryList({ categories, setter, filteredCategory, setFiltered
                 <div className=" metaDetailsBox  my-0 me-3">
                   {openEntitySelected === category.id && (
                     <div className="metaDetails ps-2 py-2">
-                      <div>ID: {category.id}</div>
-                      <div>Descripcio: {category.description}</div>
+                      {/* <div>ID: {category.id}</div> */}
+                      <div>Descripció: {category.description}</div>
                     </div>
                   )}
                 </div>
