+++ metari-app/src/components/MetaList.tsx
@@ -35,7 +35,14 @@ export function MetaList({ metas, setter, filteredCategory, groups }: MetaListPr
 
   //Si alguna de les condicions es true, es guarda la meta a la variable
   const filteredMetas = metas.filter(meta =>
-    !filteredCategory || meta.category_id === filteredCategory
+    (!filteredCategory || meta.category_id === filteredCategory) &&
+    meta.is_public &&
+    (
+      !meta.indexedMetas ||
+      meta.indexedMetas.length === 0 ||
+      meta.indexedMetas.some(im => im.is_community_approved === true) 
+      // || meta.is_public
+    )
   )
 
 
@@ -44,11 +51,11 @@ export function MetaList({ metas, setter, filteredCategory, groups }: MetaListPr
 
 
       <div className="metaList mt-4">
-        <div className="titolComponent  text-center my-2">Llista de metas</div>
-        <hr className="m-0" />
+        <div className="titolComponent  text-center my-2"><i className=" text-danger me-2 bi bi-bullseye"></i>Llistat de metas</div>
+        {/* <hr className="m-0" /> */}
 
         <div className="inline">
-          <ul className=" ps-2  m-0  py-2">
+          <ul className=" ps-3  m-0 pb-2">
             {filteredMetas.map((meta) => (
               <li key={meta.id} className="m-0 p-0" >
                 <div className={`metaEntry mt-1 me-3 ps-2 ${openEntityId === meta.id ? "mb-0" : "mb-1"} ${meta.type === "task" ? "meta-task" : "meta-challenge"}`}
@@ -59,16 +66,25 @@ export function MetaList({ metas, setter, filteredCategory, groups }: MetaListPr
 
                     {canEdit &&
                       <button className="  btn btn-warning p-1  me-2  ms-auto"
+                        title="Edita"
                         onClick={(event) => {
                           event.stopPropagation()
                           setMetaToEdit(meta)
-                        }}>Edita</button>
+                        }}><i className="bi bi-pencil"></i></button>
                     }
                     {canEdit &&
                       <button className="  btn btn-danger p-1   "
+                        title="Elimina"
                         onClick={async (event) => {
                           event.stopPropagation()
-                          await deleteMeta(meta.id)
+                          try {
+                            if (!confirm("Estàs segur que el vols eliminar?")) return;
+                            await deleteMeta(meta.id)
+                            setter(prev => prev.filter(prevMeta => prevMeta.id !== meta.id))
+                            alert("Meta eliminada correctament")
+                          } catch (error) {
+                            alert("Error en eliminar la meta")
+                          }
                         }}>X</button>
                     }
                   </div>
@@ -82,7 +98,7 @@ export function MetaList({ metas, setter, filteredCategory, groups }: MetaListPr
                         </>
                       }
                       <div>Tipus: {meta.type}</div>
-                      <div>Descripcio: {meta.description}</div>
+                      <div>Descripció: {meta.description}</div>
                       <div>Categoria: {meta.category.name}</div>
                       <div>Autor: {meta.author.username}</div>
                       <div className="d-flex mt-2 justify-content-end">
